Using parse-diff for verification
