+++ package.json
@@ -18,8 +18,8 @@
     "lint:fix": "npm run lint:fix:ts && npm run lint:fix:json",
     "lint:fix:ts": "npm run lint:ts -- --fix",
     "lint:fix:json": "prettier -w src/assets/*/*.json",
-    "verify": "ts-node scripts/verify.ts",
-    "verifyChanges": "ts-node scripts/verifyChange.ts",
+    "review:verify-deployment": "ts-node scripts/review/verifyDeployment.ts",
+    "review:diff": "ts-node scripts/review/diff.ts",
     "prepack": "npm run build",
     "test": "jest"
   },
@@ -46,6 +46,7 @@
     "eslint-plugin-prettier": "^5.1.3",
     "ethers": "^6.13.1",
     "jest": "^29.7.0",
+    "parse-diff": "^0.11.1",
     "prettier": "^3.3.2",
     "ts-jest": "^29.1.5",
     "ts-node": "^10.9.2",
